rob edits

+++ GUI/CustomModButton.xaml.cs
@@ -5,6 +5,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Windows;
+using MassSpectrometry;
+using Proteomics;
 
 namespace MetaMorpheusGUI
 {
@@ -13,43 +15,39 @@ namespace MetaMorpheusGUI
     /// </summary>
     public partial class CustomModButtonWindow : Window
     {
-        public static Dictionary<string, string> termTypeToParsableTermType;
-        public static Dictionary<string, MassSpectrometry.DissociationType> dissociationType;
+        public static Dictionary<string, string> locationRestrictions;
 
         public CustomModButtonWindow()
         {
             InitializeComponent();
 
-            if (termTypeToParsableTermType == null)
+            if (locationRestrictions == null)
             {
-                termTypeToParsableTermType = new Dictionary<string, string>();
-                termTypeToParsableTermType.Add("Anywhere", "Anywhere.");
-                termTypeToParsableTermType.Add("Peptide N-Terminus", "Peptide N-terminal.");
-                termTypeToParsableTermType.Add("Peptide C-Terminus", "Peptide C-terminal.");
-                termTypeToParsableTermType.Add("Protein N-Terminus", "N-terminal.");
-                termTypeToParsableTermType.Add("Protein C-Terminus", "C-terminal.");
+                locationRestrictions = new Dictionary<string, string>();
+                locationRestrictions.Add("Anywhere", "Anywhere.");
+                locationRestrictions.Add("Peptide N-Terminus", "Peptide N-terminal.");
+                locationRestrictions.Add("Peptide C-Terminus", "Peptide C-terminal.");
+                locationRestrictions.Add("Protein N-Terminus", "N-terminal.");
+                locationRestrictions.Add("Protein C-Terminus", "C-terminal.");
             }
 
-            foreach (var kvp in termTypeToParsableTermType)
+            foreach (string locationRestriction in locationRestrictions.Keys)
             {
-                locationRestrictionComboBox.Items.Add(kvp.Key);
+                locationRestrictionComboBox.Items.Add(locationRestriction);
             }
 
-            dissociationType = GlobalVariables.AllSupportedDissociationTypes;
-
-            foreach (var type in dissociationType.Values)
+            foreach (DissociationType type in GlobalVariables.AllSupportedDissociationTypes.Values)
             {
                 dissociationTypeComboBox.Items.Add(type);
             }
 
-            locationRestrictionComboBox.SelectedIndex = 0;
-            dissociationTypeComboBox.SelectedValue = MassSpectrometry.DissociationType.HCD;
+            locationRestrictionComboBox.SelectedItem = "Anywhere";
+            dissociationTypeComboBox.SelectedItem = DissociationType.HCD;
         }
 
         public void SaveCustomMod_Click(object sender, RoutedEventArgs e)
         {
             string modsDirectory = Path.Combine(GlobalVariables.DataDir, @"Mods");
-            var modsFiles = Directory.GetFiles(modsDirectory);
             string customModsPath = Path.Combine(modsDirectory, @"UserCustomModifications.txt");
             List<string> customModsText = new List<string>();
 
@@ -62,48 +60,87 @@ namespace MetaMorpheusGUI
                 customModsText = File.ReadAllLines(customModsPath).ToList();
             }
 
-            string myModName = originalIdTextBox.Text;
+            string idText = originalIdTextBox.Text;
             string motifText = motifTextBox.Text;
             string chemicalFormulaText = chemicalFormulaTextBox.Text;
             string modMassText = modMassTextBox.Text;
             string neutralLossText = neutralLossTextBox.Text;
             string diagnosticIonText = diagnosticIonTextBox.Text;
             string modificationTypeText = modificationTypeTextBox.Text;
-            string locationRestriction = termTypeToParsableTermType[locationRestrictionComboBox.Text];
-            MassSpectrometry.DissociationType disType = dissociationType[dissociationTypeComboBox.Text];
+            string locationRestriction = locationRestrictions[locationRestrictionComboBox.Text];
+            DissociationType disType = GlobalVariables.AllSupportedDissociationTypes[dissociationTypeComboBox.Text];
 
-            if (ErrorsDetected(myModName, motifText, modMassText, chemicalFormulaText, neutralLossText, modificationTypeText, diagnosticIonText))
+            if (ErrorsDetected(idText, motifText, modMassText, chemicalFormulaText, neutralLossText, modificationTypeText, diagnosticIonText))
             {
                 return;
             }
 
-            // write custom mod to mods file
-            customModsText.Add("ID   " + myModName);
-            customModsText.Add("TG   " + motifText);
-            customModsText.Add("PP   " + locationRestriction);
-            customModsText.Add("MT   " + modificationTypeText);
-
+            // create custom mod
+            Dictionary<DissociationType, List<double>> neutralLosses = null;
             if (!string.IsNullOrEmpty(neutralLossText))
             {
-                customModsText.Add("NL   " + neutralLossText);
+                neutralLosses = new Dictionary<DissociationType, List<double>>
+                {
+                    { disType, neutralLossText.Split(',').Select(double.Parse).ToList() }
+                };
             }
+
+            Dictionary<DissociationType, List<double>> diagnosticIons = null;
+            if (!string.IsNullOrEmpty(diagnosticIonText))
+            {
+                diagnosticIons = new Dictionary<DissociationType, List<double>>()
+                {
+                    { disType, diagnosticIonText.Split(',').Select(double.Parse).ToList() }
+                };
+            }
+
+            ModificationMotif.TryGetMotif(motifText, out ModificationMotif finalMotif);
+
+            ChemicalFormula chemicalFormula = null;
             if (!string.IsNullOrEmpty(chemicalFormulaText))
             {
-                customModsText.Add("CF   " + chemicalFormulaText);
+                chemicalFormula = ChemicalFormula.ParseFormula(chemicalFormulaText);
             }
+
+            double? modMass = null;
             if (!string.IsNullOrEmpty(modMassText))
             {
-                customModsText.Add("MM   " + modMassText);
+                modMass = double.Parse(modMassText);
             }
 
-            customModsText.Add(@"//");
+            Modification modification = new Modification(
+                _originalId: idText,
+                _modificationType: modificationTypeText,
+                _target: finalMotif,
+                _locationRestriction: locationRestriction,
+                _chemicalFormula: chemicalFormula,
+                _monoisotopicMass: modMass,
+                _neutralLosses: neutralLosses,
+                _diagnosticIons: diagnosticIons);
 
+            // write custom mod to mods file
+            
             // write/read temp file to make sure the mod is readable, then delete it
             string tempPath = Path.Combine(modsDirectory, @"temp.txt");
             try
             {
-                File.WriteAllLines(tempPath, customModsText);
-                //GlobalVariables.AddMods(UsefulProteomicsDatabases.PtmListLoader.ReadModsFromFile(tempPath));
+                List<string> temp = new List<string> { modification.ToString(), @"//" };
+                File.WriteAllLines(tempPath, temp);
+                var parsedMods = UsefulProteomicsDatabases.PtmListLoader.ReadModsFromFile(tempPath, out var errors);
+
+                if (parsedMods.Count() != 1)
+                {
+                    MessageBox.Show("Problem parsing custom mod: One mod was expected, a different number was generated", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
+                    return;
+                }
+
+                if (errors.Any())
+                {
+                    string concatErrors = string.Join(Environment.NewLine, errors.Select(p => p.Item2));
+                    MessageBox.Show("Problem(s) parsing custom mod: " + Environment.NewLine + concatErrors, "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
+                    return;
+                }
+
                 File.Delete(tempPath);
             }
             catch (Exception ex)
@@ -116,6 +153,8 @@ namespace MetaMorpheusGUI
             // delete old custom mods file, write new one
             try
             {
+                customModsText.Add(modification.ToString());
+                customModsText.Add(@"//");
                 File.Delete(customModsPath);
                 File.WriteAllLines(customModsPath, customModsText);
             }
@@ -125,51 +164,7 @@ namespace MetaMorpheusGUI
                 return;
             }
 
-            Dictionary<MassSpectrometry.DissociationType, List<double>> neutralLosses;
-            if (!string.IsNullOrEmpty(neutralLossText))
-            {
-                neutralLosses = new Dictionary<MassSpectrometry.DissociationType, List<double>>()
-            {
-                { disType, neutralLossText.Split(',').Select(double.Parse).ToList() }
-            };
-            }
-            else
-            {
-                neutralLosses = null;
-            };
-
-            Dictionary<MassSpectrometry.DissociationType, List<double>> diagnosticIons;
-            if (!string.IsNullOrEmpty(diagnosticIonText))
-            {
-                diagnosticIons = new Dictionary<MassSpectrometry.DissociationType, List<double>>()
-            {
-                { disType, diagnosticIonText.Split(',').Select(double.Parse).ToList() }
-            };
-            }
-            else
-            {
-                diagnosticIons = null;
-            }
-
-            var motif = Proteomics.ModificationMotif.TryGetMotif(motifText, out Proteomics.ModificationMotif finalMotif);
-            ChemicalFormula chemicalFormula = null;
-            if (!string.IsNullOrEmpty(chemicalFormulaText))
-            {
-                chemicalFormula = ChemicalFormula.ParseFormula(chemicalFormulaText);
-            }else
-            {
-                chemicalFormula = null;
-            }
-            double? modMass = null;
-            if (!string.IsNullOrEmpty(modMassText))
-            {
-                double.Parse(modMassText);
-            }
-
-            Proteomics.Modification newMod = new Proteomics.Modification(myModName, null, modificationTypeText, null, finalMotif, locationRestriction,
-                chemicalFormula, modMass, null, null, null, neutralLosses, diagnosticIons, null);
-
-            GlobalVariables.AddMods(new List<Proteomics.Modification> { newMod });
+            GlobalVariables.AddMods(new List<Modification> { modification }, false);
 
             DialogResult = true;
         }
@@ -184,18 +179,19 @@ namespace MetaMorpheusGUI
             // parse input
             if (string.IsNullOrEmpty(myModName) || string.IsNullOrEmpty(modType))
             {
-                MessageBox.Show("Both the mod name and type need to be specified", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
+                MessageBox.Show("The mod name and type need to be specified", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                 return true;
             }
+
             if (modType.Contains(':'))
             {
                 MessageBox.Show("Modification Type cannot contain ':'", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                 return true;
             }
+
             if (!string.IsNullOrEmpty(motif))
             {
-                var list = motif.Where(char.IsUpper);
-                if (!(list.Count() == 1))
+                if (motif.Count(char.IsUpper) != 1)
                 {
                     MessageBox.Show("Motif must contain exactly one uppercase letter", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                     return true;
@@ -206,11 +202,13 @@ namespace MetaMorpheusGUI
                 MessageBox.Show("Motif must be defined", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                 return true;
             }
+
             if (string.IsNullOrEmpty(mass) && string.IsNullOrEmpty(chemFormula))
             {
                 MessageBox.Show("Either the mass or chemical formula needs to be specified", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                 return true;
             }
+
             if (!string.IsNullOrEmpty(chemFormula))
             {
                 try
@@ -223,11 +221,13 @@ namespace MetaMorpheusGUI
                     return true;
                 }
             }
+
             if (!string.IsNullOrEmpty(mass) && !double.TryParse(mass, out double dmass))
             {
                 MessageBox.Show("Could not parse modification mass", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                 return true;
             }
+
             try
             {
                 if (!string.IsNullOrEmpty(neutralLoss))
@@ -244,6 +244,7 @@ namespace MetaMorpheusGUI
                 MessageBox.Show("Neutral losses and diagnostic ions must be entered as numbers separated by ','", "Error", MessageBoxButton.OK, MessageBoxImage.Hand);
                 return true;
             }
+
             return false;
         }
     }
